Integrate SOTA loss strategies and enhance medical classification pipeline
