--- ENHANCED_FINETUNE_README.md
@@ -1,191 +1,0 @@
-# Enhanced SeqSetVAE Finetuning for Medical Classification
-
-## 🎯 Objective
-Improve classification performance from current metrics to target metrics:
-- **AUC**: 0.7285 → **≥0.90**
-- **AUPRC**: 0.2903 → **≥0.50**
-- **Accuracy**: 0.5484 → **≥0.70**
-
-## 🚀 Key Improvements Implemented
-
-### 1. **Advanced Classification Head Architecture**
-- **Multi-head self-attention** for feature refinement
-- **Dual-pathway processing** with gated feature fusion
-- **Residual connections** and layer normalization
-- **Auxiliary prediction head** for better gradient flow
-- **Enhanced dropout** strategies for better generalization
-
-### 2. **Improved Loss Function Strategy**
-- **Enhanced Focal Loss** with adaptive class balancing
-- **Auxiliary loss** (30% weight) for better training dynamics
-- **Optimized focal parameters**: α=0.20, γ=2.5
-- **Reduced label smoothing** for sharper boundaries
-
-### 3. **Optimized Training Hyperparameters**
-- **Higher learning rates** for complex architecture (8e-4)
-- **Differentiated LR** for attention vs classifier components
-- **Cosine annealing with warm restarts**
-- **Enhanced gradient clipping** (0.3)
-- **Increased training epochs** (25)
-
-### 4. **Advanced Optimization Features**
-- **Parameter group optimization** with different LR schedules
-- **Enhanced early stopping** with AUC monitoring
-- **Adaptive focal loss** computed from training data
-- **Step-wise LR scheduling** for fine-grained control
-
-## 📋 Usage Instructions
-
-### Quick Start
-```bash
-# Run enhanced finetuning
-python finetune_enhanced.py \
-    --pretrained_ckpt /path/to/pretrained/checkpoint.ckpt \
-    --data_dir /path/to/data \
-    --output_dir /path/to/outputs
-```
-
-### Full Command Example
-```bash
-python finetune_enhanced.py \
-    --pretrained_ckpt /home/sunx/data/aiiih/projects/sunx/projects/SeqSetVAE/outputs/SeqSetVAE-PT/checkpoints/SeqSetVAE_pretrain_batch4.ckpt \
-    --data_dir /home/sunx/data/aiiih/data/mimic/processed/patient_ehr \
-    --params_map_path /home/sunx/data/aiiih/data/mimic/processed/stats.csv \
-    --label_path /home/sunx/data/aiiih/data/mimic/processed/oc.csv \
-    --output_dir /home/sunx/data/aiiih/projects/sunx/projects/TEEMR/PT/outputs_enhanced \
-    --seed 42 \
-    --devices 1
-```
-
-### Alternative: Use Original Training Script with Enhanced Config
-```bash
-python train.py \
-    --mode finetune \
-    --pretrained_ckpt /path/to/pretrained.ckpt \
-    --batch_size 6 \
-    --max_epochs 25 \
-    --data_dir /path/to/data \
-    --output_dir /path/to/outputs
-```
-
-## 📊 Expected Improvements
-
-### Training Dynamics
-- **Faster convergence** due to higher learning rates
-- **Better gradient flow** from auxiliary loss
-- **More stable training** from enhanced architecture
-- **Improved generalization** from advanced regularization
-
-### Performance Metrics
-- **AUC improvement**: +0.17 (0.73 → 0.90)
-- **AUPRC improvement**: +0.20 (0.29 → 0.50+)
-- **Better precision-recall balance**
-- **More robust predictions**
-
-## 🔧 Configuration Files
-
-### Enhanced Configuration (`finetune_config.py`)
-Key improvements:
-- `cls_head_lr = 8e-4` (increased from 3e-4)
-- `focal_alpha = 0.20` (optimized from 0.25)
-- `focal_gamma = 2.5` (increased from 2.0)
-- `max_epochs = 25` (increased from 15)
-- `early_stopping_patience = 8` (increased from 6)
-
-### Model Architecture Changes
-- **Advanced classifier head** with attention mechanisms
-- **Auxiliary prediction path** for better training
-- **Enhanced parameter initialization**
-- **Optimized dropout strategies**
-
-## 📈 Monitoring and Validation
-
-### Real-time Metrics
-- **AUC/AUPRC progress** tracked every epoch
-- **Combined score** (0.6×AUC + 0.4×AUPRC)
-- **Best metrics** displayed during training
-- **Learning rate monitoring**
-
-### Checkpointing Strategy
-- **Top-3 checkpoints** saved based on AUC
-- **Automatic filename** with metrics
-- **Last checkpoint** always saved
-- **Enhanced model selection**
-
-## 🎛️ Advanced Features
-
-### Adaptive Training
-- **Automatic focal loss balancing** from data statistics
-- **Dynamic learning rate scheduling**
-- **Gradient accumulation** for stability
-- **Memory-efficient training**
-
-### Debugging and Analysis
-- **Posterior collapse monitoring**
-- **Gradient norm tracking**
-- **Detailed parameter counting**
-- **Training progress visualization**
-
-## 🔬 Technical Details
-
-### Architecture Enhancements
-```python
-# Advanced classifier components:
-- Feature enhancer (Linear + LayerNorm + GELU)
-- Self-attention (8 heads, dropout=0.1)
-- Dual pathways (GELU + ReLU)
-- Gated fusion mechanism
-- Auxiliary prediction head
-```
-
-### Loss Function
-```python
-# Enhanced loss computation:
-main_loss = focal_loss(main_logits, labels)
-aux_loss = focal_loss(aux_logits, labels)
-total_loss = main_loss + 0.3 * aux_loss
-```
-
-### Optimizer Configuration
-```python
-# Differentiated learning rates:
-attention_params: cls_lr * 0.8
-classifier_params: cls_lr * 1.0
-scheduler: CosineAnnealingWarmRestarts
-```
-
-## 🚨 Troubleshooting
-
-### Common Issues
-1. **CUDA memory errors**: Reduce batch_size to 4
-2. **Slow convergence**: Increase cls_head_lr to 1e-3
-3. **Overfitting**: Increase dropout to 0.3
-4. **Unstable training**: Reduce gradient_clip_val to 0.2
-
-### Performance Tips
-- **Monitor combined score** rather than individual metrics
-- **Allow sufficient training time** (20+ epochs)
-- **Use GPU** for optimal performance
-- **Set deterministic=True** for reproducible results
-
-## 📝 Expected Training Output
-```
-🚀 Enhanced SeqSetVAE Finetuning Configuration:
- - Max epochs: 25
- - Classification head LR: 0.0008
- - Focal loss: α=0.20, γ=2.5
-
-🧊 Enhanced Finetuning Configuration:
- - Trainable parameters: ~200K
- - Advanced classifier: Multi-head attention + residual connections
- - Auxiliary loss: 30% weight for better gradient flow
-
-🎯 Current: AUC=0.8234, AUPRC=0.4567, Combined=0.6763
-🏆 Best: AUC=0.8234, AUPRC=0.4567, Combined=0.6763
-
-✅ AUC target achieved!
-📈 AUPRC progress: 0.4567/0.50
-```
-
-## 📞 Support
-If you encounter issues or need further optimization, the enhanced architecture provides multiple tuning points for different medical datasets and class distributions.
