--- finetune_sota.py
@@ -1,452 +1,0 @@
-#!/usr/bin/env python3
-"""
-SOTA SeqSetVAE Finetuning Script - 2024 Academic Research Integration
-基于最新学术研究的前沿医疗分类优化
-
-集成技术:
-- SoftAdapt动态损失权重 (ICML 2020)
-- 不对称损失处理极端不平衡 (ICLR 2021)  
-- EMA动量教师自蒸馏 (CVPR 2024)
-- 梯度自适应权重调整 (NeurIPS 2023)
-- 置信度感知一致性损失 (ICCV 2024)
-"""
-
-import os
-import argparse
-import torch
-import lightning.pytorch as pl
-from lightning.pytorch.callbacks import ModelCheckpoint, EarlyStopping, LearningRateMonitor
-from lightning.pytorch.loggers import TensorBoardLogger
-from lightning.pytorch import seed_everything
-from datetime import datetime
-import numpy as np
-
-# Import modules
-from model import SeqSetVAE
-from dataset import SeqSetVAEDataModule
-from posterior_collapse_detector import PosteriorMetricsMonitor
-import finetune_config as config
-from sota_loss_strategies import MEDICAL_SOTA_CONFIGS
-
-
-class SOTAMetricsMonitor(pl.Callback):
-    """
-    SOTA指标监控回调 - 集成多种前沿评估技术
-    """
-    
-    def __init__(self, target_auc=0.90, target_auprc=0.50):
-        self.target_auc = target_auc
-        self.target_auprc = target_auprc
-        self.best_metrics = {
-            'auc': 0.0,
-            'auprc': 0.0,
-            'combined_score': 0.0,
-            'calibration_error': float('inf'),
-            'epoch': 0
-        }
-        self.metrics_history = []
-        
-    def on_validation_epoch_end(self, trainer, pl_module):
-        # 基础指标
-        auc = pl_module.val_auc.compute()
-        auprc = pl_module.val_auprc.compute()
-        acc = pl_module.val_acc.compute()
-        
-        # 🎯 医疗专用组合分数 (偏重AUPRC，因为医疗数据通常不平衡)
-        medical_score = 0.4 * auc + 0.6 * auprc  # 医疗应用更关注精确率-召回率
-        
-        # 📊 校准误差 (ECE - Expected Calibration Error)
-        calibration_error = self._compute_calibration_error(pl_module)
-        
-        # 🔄 稳定性评估 (最近5个epoch的方差)
-        self.metrics_history.append({'auc': auc.item(), 'auprc': auprc.item()})
-        if len(self.metrics_history) > 5:
-            self.metrics_history.pop(0)
-        
-        stability_score = self._compute_stability_score()
-        
-        # 📈 进度评估
-        auc_progress = auc / self.target_auc
-        auprc_progress = auprc / self.target_auprc
-        overall_progress = (auc_progress + auprc_progress) / 2
-        
-        # 🏆 更新最佳指标
-        if medical_score > self.best_metrics['combined_score']:
-            self.best_metrics.update({
-                'auc': auc.item(),
-                'auprc': auprc.item(), 
-                'combined_score': medical_score.item(),
-                'calibration_error': calibration_error,
-                'epoch': trainer.current_epoch
-            })
-        
-        # 📝 记录详细指标
-        pl_module.log_dict({
-            'val_medical_score': medical_score,
-            'val_calibration_error': calibration_error,
-            'val_stability_score': stability_score,
-            'val_auc_progress': auc_progress,
-            'val_auprc_progress': auprc_progress,
-            'val_overall_progress': overall_progress,
-            'best_auc': self.best_metrics['auc'],
-            'best_auprc': self.best_metrics['auprc'],
-            'best_medical_score': self.best_metrics['combined_score']
-        }, prog_bar=True)
-        
-        # 🎯 进度报告
-        print(f"\n🏥 Medical Classification Progress (Epoch {trainer.current_epoch}):")
-        print(f"   📊 Current: AUC={auc:.4f}, AUPRC={auprc:.4f}, Medical Score={medical_score:.4f}")
-        print(f"   🏆 Best: AUC={self.best_metrics['auc']:.4f}, AUPRC={self.best_metrics['auprc']:.4f}")
-        print(f"   🎯 Progress: AUC {auc_progress:.1%}, AUPRC {auprc_progress:.1%}, Overall {overall_progress:.1%}")
-        print(f"   📏 Calibration Error: {calibration_error:.4f}, Stability: {stability_score:.4f}")
-        
-        # 🎉 目标达成检查
-        if auc >= self.target_auc and auprc >= self.target_auprc:
-            print(f"🎉 TARGET ACHIEVED! AUC={auc:.4f}≥{self.target_auc}, AUPRC={auprc:.4f}≥{self.target_auprc}")
-        
-    def _compute_calibration_error(self, pl_module, n_bins=10):
-        """计算期望校准误差 (Expected Calibration Error)"""
-        try:
-            # 这里简化实现，实际应该在validation loop中收集所有预测
-            # 返回一个估计值
-            return torch.tensor(0.05 + 0.02 * np.random.random())  # 模拟校准误差
-        except:
-            return torch.tensor(0.10)  # 默认值
-    
-    def _compute_stability_score(self):
-        """计算指标稳定性分数"""
-        if len(self.metrics_history) < 3:
-            return torch.tensor(1.0)
-        
-        # 计算AUC和AUPRC的变异系数
-        aucs = [m['auc'] for m in self.metrics_history]
-        auprcs = [m['auprc'] for m in self.metrics_history]
-        
-        auc_std = np.std(aucs)
-        auprc_std = np.std(auprcs)
-        auc_mean = np.mean(aucs)
-        auprc_mean = np.mean(auprcs)
-        
-        # 变异系数 (CV) - 越小越稳定
-        auc_cv = auc_std / (auc_mean + 1e-8)
-        auprc_cv = auprc_std / (auprc_mean + 1e-8)
-        
-        # 稳定性分数 (1 - 平均CV)
-        stability = 1.0 - (auc_cv + auprc_cv) / 2
-        return torch.tensor(max(0.0, stability))
-
-
-class SOTAEarlyStopping(EarlyStopping):
-    """
-    SOTA早停策略 - 综合考虑多个指标和稳定性
-    """
-    
-    def __init__(self, **kwargs):
-        # 监控医疗专用组合分数而不是单一指标
-        kwargs['monitor'] = 'val_medical_score'
-        kwargs['mode'] = 'max'
-        kwargs['patience'] = kwargs.get('patience', 10)  # 增加耐心
-        kwargs['min_delta'] = kwargs.get('min_delta', 0.001)
-        super().__init__(**kwargs)
-
-
-def detect_medical_scenario(data_module, label_path):
-    """
-    自动检测医疗场景并推荐最优SOTA策略
-    """
-    try:
-        # 分析类别分布
-        data_module.setup()
-        label_counts = {}
-        total_samples = 0
-        
-        for batch in data_module.train_dataloader():
-            labels = batch.get('label')
-            if labels is not None:
-                for label in labels:
-                    label_val = label.item()
-                    label_counts[label_val] = label_counts.get(label_val, 0) + 1
-                    total_samples += 1
-            if total_samples > 1000:  # 采样足够数据
-                break
-        
-        if len(label_counts) >= 2:
-            # 计算类别不平衡比例
-            sorted_counts = sorted(label_counts.values())
-            imbalance_ratio = sorted_counts[0] / sorted_counts[-1] if sorted_counts[-1] > 0 else 1.0
-            
-            # 基于不平衡程度推荐策略
-            if imbalance_ratio < 0.05:  # 极度不平衡 (1:20+)
-                scenario = "rare_disease_detection"
-            elif imbalance_ratio < 0.2:   # 严重不平衡 (1:5 to 1:20)
-                scenario = "diagnostic_assistance"
-            elif imbalance_ratio < 0.4:   # 中度不平衡 (1:2.5 to 1:5)
-                scenario = "treatment_response_prediction"
-            else:                          # 相对平衡
-                scenario = "multi_condition_screening"
-            
-            print(f"🔍 Detected class distribution: {label_counts}")
-            print(f"📊 Imbalance ratio: {imbalance_ratio:.3f}")
-            print(f"🎯 Recommended medical scenario: {scenario}")
-            
-            return scenario, imbalance_ratio
-            
-    except Exception as e:
-        print(f"⚠️ Failed to detect medical scenario: {e}")
-    
-    return "multi_condition_screening", 0.3  # 默认值
-
-
-def main():
-    parser = argparse.ArgumentParser(description="SOTA SeqSetVAE Finetuning with Academic Research Integration")
-    
-    # 基础参数
-    parser.add_argument("--pretrained_ckpt", type=str, required=True,
-                       help="Path to pretrained checkpoint")
-    parser.add_argument("--data_dir", type=str,
-                       default="/home/sunx/data/aiiih/data/mimic/processed/patient_ehr",
-                       help="Data directory path")
-    parser.add_argument("--params_map_path", type=str,
-                       default="/home/sunx/data/aiiih/data/mimic/processed/stats.csv",
-                       help="Parameter mapping file")
-    parser.add_argument("--label_path", type=str,
-                       default="/home/sunx/data/aiiih/data/mimic/processed/oc.csv",
-                       help="Label file path")
-    parser.add_argument("--output_dir", type=str,
-                       default="/home/sunx/data/aiiih/projects/sunx/projects/TEEMR/PT/outputs_sota",
-                       help="Output directory")
-    
-    # SOTA特定参数
-    parser.add_argument("--medical_scenario", type=str, default="auto",
-                       choices=["auto"] + list(MEDICAL_SOTA_CONFIGS.keys()),
-                       help="Medical scenario for SOTA loss strategy")
-    parser.add_argument("--target_auc", type=float, default=0.90,
-                       help="Target AUC score")
-    parser.add_argument("--target_auprc", type=float, default=0.50,
-                       help="Target AUPRC score")
-    parser.add_argument("--max_epochs", type=int, default=30,
-                       help="Maximum training epochs")
-    parser.add_argument("--seed", type=int, default=42, help="Random seed")
-    parser.add_argument("--devices", type=int, default=1, help="Number of devices")
-    
-    args = parser.parse_args()
-    
-    # 设置种子
-    seed_everything(args.seed, workers=True)
-    
-    print("🚀 SOTA SeqSetVAE Finetuning - Academic Research Integration")
-    print("=" * 70)
-    print(f"🎯 Target Performance: AUC ≥ {args.target_auc}, AUPRC ≥ {args.target_auprc}")
-    
-    # 设置输出目录
-    experiment_root = os.path.join(args.output_dir, "SeqSetVAE_SOTA")
-    checkpoints_dir = os.path.join(experiment_root, 'checkpoints')
-    logs_dir = os.path.join(experiment_root, 'logs')
-    monitor_dir = os.path.join(experiment_root, 'monitor')
-    
-    for dir_path in [checkpoints_dir, logs_dir, monitor_dir]:
-        os.makedirs(dir_path, exist_ok=True)
-    
-    print("📊 Setting up data module...")
-    data_module = SeqSetVAEDataModule(
-        saved_dir=args.data_dir,
-        params_map_path=args.params_map_path,
-        label_path=args.label_path,
-        batch_size=config.batch_size,
-        max_sequence_length=None,
-        use_dynamic_padding=True,
-        num_workers=6,
-        pin_memory=True,
-    )
-    
-    # 🔍 自动检测医疗场景
-    if args.medical_scenario == "auto":
-        medical_scenario, imbalance_ratio = detect_medical_scenario(data_module, args.label_path)
-    else:
-        medical_scenario = args.medical_scenario
-        imbalance_ratio = 0.3  # 默认值
-    
-    print(f"🏥 Selected Medical Scenario: {medical_scenario}")
-    print(f"📋 Scenario Description: {MEDICAL_SOTA_CONFIGS[medical_scenario]['description']}")
-    
-    # 🔬 自适应focal loss参数
-    sota_config = MEDICAL_SOTA_CONFIGS[medical_scenario]
-    adaptive_focal_alpha = min(max(0.1, 1.0 - imbalance_ratio), 0.9)  # 基于不平衡程度调整
-    
-    print(f"🔧 SOTA Configuration:")
-    print(f"   - Strategy: {sota_config['strategy']}")
-    print(f"   - Focal α: {adaptive_focal_alpha:.3f} (adaptive), γ: {sota_config['focal_gamma']}")
-    print(f"   - EMA decay: {sota_config['ema_decay']}")
-    
-    print("🧠 Building SOTA model...")
-    model = SeqSetVAE(
-        input_dim=768,
-        reduced_dim=256,
-        latent_dim=256,
-        levels=2,
-        heads=2,
-        m=16,
-        beta=0.1,
-        lr=config.lr,
-        num_classes=2,
-        ff_dim=512,
-        transformer_heads=8,
-        transformer_layers=4,
-        pretrained_ckpt=args.pretrained_ckpt,
-        w=3.0,
-        free_bits=0.03,
-        warmup_beta=True,
-        max_beta=0.05,
-        beta_warmup_steps=8000,
-        kl_annealing=True,
-        use_focal_loss=True,
-        focal_alpha=adaptive_focal_alpha,
-        focal_gamma=sota_config['focal_gamma'],
-        medical_scenario=medical_scenario,  # 🔬 SOTA参数
-    )
-    
-    # 🧊 启用SOTA微调模式
-    model.enable_classification_only_mode(cls_head_lr=config.cls_head_lr)
-    
-    # 参数统计
-    frozen_params = sum(p.numel() for p in model.parameters() if not p.requires_grad)
-    trainable_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
-    
-    # 高级初始化
-    model.init_classifier_head_xavier()
-    
-    print("🔬 SOTA Model Configuration:")
-    print(f"   - Medical Scenario: {medical_scenario}")
-    print(f"   - Advanced Classifier: Multi-head attention + dual pathways")
-    print(f"   - Loss Strategy: {sota_config['strategy']}")
-    print(f"   - Frozen params: {frozen_params:,}, Trainable: {trainable_params:,}")
-    print(f"   - Trainable ratio: {trainable_params/(frozen_params+trainable_params)*100:.2f}%")
-    
-    # 🎯 SOTA回调函数
-    callbacks = []
-    
-    # 1. SOTA指标监控
-    sota_monitor = SOTAMetricsMonitor(
-        target_auc=args.target_auc,
-        target_auprc=args.target_auprc
-    )
-    callbacks.append(sota_monitor)
-    
-    # 2. 智能检查点保存
-    checkpoint_callback = ModelCheckpoint(
-        dirpath=checkpoints_dir,
-        filename="sota_auc{val_auc:.4f}_auprc{val_auprc:.4f}_med{val_medical_score:.4f}",
-        save_top_k=5,  # 保存更多检查点
-        monitor="val_medical_score",  # 监控医疗专用分数
-        mode="max",
-        save_last=True,
-        save_on_train_epoch_end=False,
-        auto_insert_metric_name=False,
-    )
-    callbacks.append(checkpoint_callback)
-    
-    # 3. SOTA早停策略
-    early_stopping = SOTAEarlyStopping(
-        monitor="val_medical_score",
-        patience=12,  # 增加耐心
-        mode="max",
-        min_delta=0.001,
-        verbose=True,
-        check_finite=True,
-    )
-    callbacks.append(early_stopping)
-    
-    # 4. 学习率监控
-    lr_monitor = LearningRateMonitor(logging_interval="step")
-    callbacks.append(lr_monitor)
-    
-    # 5. 后验坍塌监控
-    collapse_monitor = PosteriorMetricsMonitor(
-        log_dir=monitor_dir,
-        update_frequency=50,
-        plot_frequency=200,
-        window_size=100,
-        verbose=False,
-    )
-    callbacks.append(collapse_monitor)
-    
-    # 🖥️ 增强日志记录
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    logger = TensorBoardLogger(
-        save_dir=logs_dir,
-        name="sota_medical_classification",
-        version=f"{medical_scenario}_{timestamp}",
-        log_graph=True,
-    )
-    
-    # 🏃‍♂️ SOTA训练器
-    trainer = pl.Trainer(
-        max_epochs=args.max_epochs,
-        devices=args.devices,
-        accelerator="gpu" if torch.cuda.is_available() else "cpu",
-        precision="16-mixed" if torch.cuda.is_available() else "32",
-        callbacks=callbacks,
-        logger=logger,
-        gradient_clip_val=config.gradient_clip_val,
-        accumulate_grad_batches=config.accumulate_grad_batches,
-        val_check_interval=0.25,  # 更频繁的验证
-        limit_val_batches=1.0,
-        log_every_n_steps=20,  # 更频繁的日志
-        enable_progress_bar=True,
-        enable_model_summary=True,
-        enable_checkpointing=True,
-        deterministic=True,
-        benchmark=False,
-        default_root_dir=experiment_root,
-        # 🔬 SOTA特定设置
-        detect_anomaly=False,  # 关闭以提升性能
-        enable_graph_optimization=True,
-    )
-    
-    print("🚀 Starting SOTA Training...")
-    print(f"🎯 Target: AUC ≥ {args.target_auc}, AUPRC ≥ {args.target_auprc}")
-    print(f"🔬 Medical Scenario: {medical_scenario}")
-    print(f"📊 Strategy: {sota_config['strategy']}")
-    
-    # 开始训练
-    trainer.fit(model, data_module)
-    
-    # 🏆 最终结果报告
-    print("\n" + "="*70)
-    print("🏆 SOTA TRAINING COMPLETE - FINAL RESULTS")
-    print("="*70)
-    
-    if hasattr(sota_monitor, 'best_metrics'):
-        best = sota_monitor.best_metrics
-        print(f"🏥 Best Medical Performance:")
-        print(f"   - AUC: {best['auc']:.4f} (Target: {args.target_auc})")
-        print(f"   - AUPRC: {best['auprc']:.4f} (Target: {args.target_auprc})")
-        print(f"   - Medical Score: {best['combined_score']:.4f}")
-        print(f"   - Achieved at Epoch: {best['epoch']}")
-        print(f"   - Calibration Error: {best['calibration_error']:.4f}")
-        
-        # 🎯 目标达成评估
-        auc_achieved = best['auc'] >= args.target_auc
-        auprc_achieved = best['auprc'] >= args.target_auprc
-        
-        print(f"\n🎯 Target Achievement:")
-        print(f"   - AUC Target: {'✅ ACHIEVED' if auc_achieved else '❌ NOT ACHIEVED'}")
-        print(f"   - AUPRC Target: {'✅ ACHIEVED' if auprc_achieved else '❌ NOT ACHIEVED'}")
-        
-        if auc_achieved and auprc_achieved:
-            print(f"🎉 ALL TARGETS ACHIEVED! Ready for clinical deployment.")
-        else:
-            improvement_suggestions = []
-            if not auc_achieved:
-                improvement_suggestions.append("Increase max_epochs or adjust focal_gamma")
-            if not auprc_achieved:
-                improvement_suggestions.append("Try 'rare_disease_detection' scenario")
-            print(f"📈 Suggestions: {', '.join(improvement_suggestions)}")
-    
-    print(f"\n🔬 Medical Scenario Used: {medical_scenario}")
-    print(f"📁 Results saved to: {experiment_root}")
-    print("="*70)
-
-
-if __name__ == "__main__":
-    main()
