--- sota_loss_strategies.py
@@ -1,471 +1,0 @@
-"""
-State-of-the-Art Loss Strategies for Medical Classification (2024)
-基于最新学术研究的前沿损失函数策略
-
-References:
-- SoftAdapt: Dynamic Loss Weighting (ICML 2020)
-- Asymmetric Loss for Multi-Label Classification (ICLR 2021)
-- Self-Distillation with Momentum Teacher (CVPR 2024)
-- Gradient-based Auxiliary Loss Adaptation (NeurIPS 2023)
-"""
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-import numpy as np
-from typing import Dict, Tuple, Optional
-from losses import FocalLoss
-
-
-class SOTALossStrategy:
-    """
-    State-of-the-art Loss Strategy combining multiple recent advances:
-    1. SoftAdapt for dynamic loss weighting
-    2. Asymmetric Loss for imbalanced classification
-    3. Self-Distillation with EMA teacher
-    4. Gradient-based adaptation
-    """
-    
-    def __init__(
-        self,
-        num_classes: int = 2,
-        focal_alpha: float = 0.2,
-        focal_gamma: float = 2.5,
-        ema_decay: float = 0.999,
-        adaptation_lr: float = 0.01,
-        strategy: str = "adaptive_distillation"
-    ):
-        self.num_classes = num_classes
-        self.strategy = strategy
-        self.ema_decay = ema_decay
-        self.adaptation_lr = adaptation_lr
-        
-        # 1. Advanced Focal Loss for main task
-        self.focal_loss = FocalLoss(alpha=focal_alpha, gamma=focal_gamma, reduction="mean")
-        
-        # 2. Asymmetric Loss for auxiliary task (handles extreme imbalance)
-        self.asymmetric_loss = AsymmetricLoss(gamma_neg=4, gamma_pos=1, clip=0.05)
-        
-        # 3. SoftAdapt weights (learnable parameters)
-        self.register_loss_weights()
-        
-        # 4. EMA teacher for self-distillation
-        self.ema_teacher_logits = None
-        
-        # 5. Gradient similarity tracking
-        self.prev_main_grad = None
-        self.prev_aux_grad = None
-        
-    def register_loss_weights(self):
-        """Register learnable loss weights (SoftAdapt approach)"""
-        # Initialize as learnable parameters
-        self.main_weight = nn.Parameter(torch.tensor(1.0))
-        self.aux_weight = nn.Parameter(torch.tensor(0.5))
-        self.distill_weight = nn.Parameter(torch.tensor(0.3))
-        self.consistency_weight = nn.Parameter(torch.tensor(0.1))
-        
-    def compute_loss(
-        self,
-        main_logits: torch.Tensor,
-        aux_logits: torch.Tensor,
-        labels: torch.Tensor,
-        epoch: int = 0,
-        main_params: Optional[list] = None,
-        aux_params: Optional[list] = None
-    ) -> Tuple[torch.Tensor, Dict[str, torch.Tensor]]:
-        """
-        Compute SOTA loss with multiple advanced techniques
-        
-        Args:
-            main_logits: Main head predictions [B, C]
-            aux_logits: Auxiliary head predictions [B, C]
-            labels: Ground truth labels [B]
-            epoch: Current epoch for scheduling
-            main_params: Main head parameters for gradient analysis
-            aux_params: Auxiliary head parameters for gradient analysis
-        """
-        
-        device = main_logits.device
-        
-        # 1. Main Loss: Advanced Focal Loss
-        main_loss = self.focal_loss(main_logits, labels)
-        
-        # 2. Auxiliary Loss: Asymmetric Loss (better for extreme imbalance)
-        aux_loss = self.asymmetric_loss(aux_logits, labels)
-        
-        # 3. Self-Distillation Loss with EMA Teacher
-        distill_loss = self._compute_distillation_loss(main_logits, aux_logits, epoch)
-        
-        # 4. Advanced Consistency Loss
-        consistency_loss = self._compute_consistency_loss(main_logits, aux_logits, labels)
-        
-        # 5. Dynamic Weight Adaptation (SoftAdapt + Gradient-based)
-        adapted_weights = self._adapt_loss_weights(
-            main_loss, aux_loss, distill_loss, consistency_loss,
-            main_params, aux_params
-        )
-        
-        # 6. Combine losses with adapted weights
-        total_loss = (
-            adapted_weights['main'] * main_loss +
-            adapted_weights['aux'] * aux_loss +
-            adapted_weights['distill'] * distill_loss +
-            adapted_weights['consistency'] * consistency_loss
-        )
-        
-        # Loss breakdown for monitoring
-        loss_dict = {
-            'main_loss': main_loss,
-            'aux_loss': aux_loss,
-            'distill_loss': distill_loss,
-            'consistency_loss': consistency_loss,
-            'main_weight': adapted_weights['main'],
-            'aux_weight': adapted_weights['aux'],
-            'distill_weight': adapted_weights['distill'],
-            'consistency_weight': adapted_weights['consistency']
-        }
-        
-        return total_loss, loss_dict
-    
-    def _compute_distillation_loss(
-        self, 
-        main_logits: torch.Tensor, 
-        aux_logits: torch.Tensor, 
-        epoch: int
-    ) -> torch.Tensor:
-        """
-        Self-distillation with EMA teacher (inspired by momentum contrast)
-        """
-        # Update EMA teacher with main head predictions
-        main_probs = F.softmax(main_logits.detach(), dim=1)
-        
-        if self.ema_teacher_logits is None:
-            self.ema_teacher_logits = main_probs.clone()
-        else:
-            self.ema_teacher_logits = (
-                self.ema_decay * self.ema_teacher_logits + 
-                (1 - self.ema_decay) * main_probs
-            )
-        
-        # Knowledge distillation from EMA teacher to auxiliary head
-        temperature = 3.0 + 2.0 * np.exp(-epoch / 10)  # Adaptive temperature
-        
-        aux_log_probs = F.log_softmax(aux_logits / temperature, dim=1)
-        teacher_probs = F.softmax(self.ema_teacher_logits / temperature, dim=1)
-        
-        distill_loss = F.kl_div(aux_log_probs, teacher_probs, reduction='batchmean')
-        
-        return distill_loss * (temperature ** 2)  # Scale by temperature^2
-    
-    def _compute_consistency_loss(
-        self,
-        main_logits: torch.Tensor,
-        aux_logits: torch.Tensor, 
-        labels: torch.Tensor
-    ) -> torch.Tensor:
-        """
-        Advanced consistency loss with uncertainty weighting
-        """
-        main_probs = F.softmax(main_logits, dim=1)
-        aux_probs = F.softmax(aux_logits, dim=1)
-        
-        # Confidence-based weighting (higher weight for confident predictions)
-        main_confidence = torch.max(main_probs, dim=1)[0]
-        aux_confidence = torch.max(aux_probs, dim=1)[0]
-        
-        # Agreement-based consistency (only for samples where both are confident)
-        confidence_mask = (main_confidence > 0.7) & (aux_confidence > 0.7)
-        
-        if confidence_mask.sum() > 0:
-            # Bidirectional KL divergence for confident samples
-            kl_main_to_aux = F.kl_div(
-                F.log_softmax(aux_logits[confidence_mask], dim=1),
-                main_probs[confidence_mask].detach(),
-                reduction='batchmean'
-            )
-            kl_aux_to_main = F.kl_div(
-                F.log_softmax(main_logits[confidence_mask], dim=1),
-                aux_probs[confidence_mask].detach(),
-                reduction='batchmean'
-            )
-            consistency_loss = 0.5 * (kl_main_to_aux + kl_aux_to_main)
-        else:
-            consistency_loss = torch.tensor(0.0, device=main_logits.device)
-            
-        return consistency_loss
-    
-    def _adapt_loss_weights(
-        self,
-        main_loss: torch.Tensor,
-        aux_loss: torch.Tensor,
-        distill_loss: torch.Tensor,
-        consistency_loss: torch.Tensor,
-        main_params: Optional[list] = None,
-        aux_params: Optional[list] = None
-    ) -> Dict[str, torch.Tensor]:
-        """
-        SoftAdapt + Gradient-based adaptive weighting
-        """
-        # Base weights (learnable parameters)
-        base_weights = {
-            'main': torch.sigmoid(self.main_weight),
-            'aux': torch.sigmoid(self.aux_weight), 
-            'distill': torch.sigmoid(self.distill_weight),
-            'consistency': torch.sigmoid(self.consistency_weight)
-        }
-        
-        # Gradient-based adaptation (if parameters provided)
-        if main_params is not None and aux_params is not None:
-            grad_adaptation = self._compute_gradient_adaptation(
-                main_loss, aux_loss, main_params, aux_params
-            )
-            
-            # Combine base weights with gradient adaptation
-            adapted_weights = {}
-            for key in base_weights:
-                if key in grad_adaptation:
-                    adapted_weights[key] = base_weights[key] * grad_adaptation[key]
-                else:
-                    adapted_weights[key] = base_weights[key]
-        else:
-            adapted_weights = base_weights
-        
-        # Normalize weights to sum to reasonable range
-        total_weight = sum(adapted_weights.values())
-        normalized_weights = {k: v / total_weight * 2.0 for k, v in adapted_weights.items()}
-        
-        return normalized_weights
-    
-    def _compute_gradient_adaptation(
-        self,
-        main_loss: torch.Tensor,
-        aux_loss: torch.Tensor,
-        main_params: list,
-        aux_params: list
-    ) -> Dict[str, torch.Tensor]:
-        """
-        Gradient-based loss weight adaptation
-        """
-        # Compute gradients
-        main_grads = torch.autograd.grad(main_loss, main_params, retain_graph=True, allow_unused=True)
-        aux_grads = torch.autograd.grad(aux_loss, aux_params, retain_graph=True, allow_unused=True)
-        
-        # Flatten gradients
-        main_grad_flat = torch.cat([g.flatten() for g in main_grads if g is not None])
-        aux_grad_flat = torch.cat([g.flatten() for g in aux_grads if g is not None])
-        
-        if len(main_grad_flat) == 0 or len(aux_grad_flat) == 0:
-            return {'main': 1.0, 'aux': 1.0}
-        
-        # Compute gradient similarity (cosine similarity)
-        similarity = F.cosine_similarity(
-            main_grad_flat.unsqueeze(0), 
-            aux_grad_flat.unsqueeze(0)
-        ).item()
-        
-        # Adaptive weighting based on gradient similarity
-        # High similarity -> reduce auxiliary weight (avoid redundancy)
-        # Low similarity -> increase auxiliary weight (complementary learning)
-        main_adapt = 1.0
-        aux_adapt = max(0.1, 1.0 - abs(similarity))  # Inverse relationship
-        
-        return {
-            'main': torch.tensor(main_adapt, device=main_loss.device),
-            'aux': torch.tensor(aux_adapt, device=main_loss.device)
-        }
-
-
-class AsymmetricLoss(nn.Module):
-    """
-    Asymmetric Loss for Multi-Label Classification (ICLR 2021)
-    Optimized for extreme class imbalance in medical data
-    """
-    
-    def __init__(
-        self,
-        gamma_neg: float = 4,
-        gamma_pos: float = 1,
-        clip: float = 0.05,
-        eps: float = 1e-8,
-        disable_torch_grad_focal_loss: bool = True
-    ):
-        super().__init__()
-        self.gamma_neg = gamma_neg
-        self.gamma_pos = gamma_pos
-        self.clip = clip
-        self.disable_torch_grad_focal_loss = disable_torch_grad_focal_loss
-        self.eps = eps
-        
-    def forward(self, x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
-        """
-        Forward pass for asymmetric loss
-        
-        Args:
-            x: Logits [B, C]
-            y: Labels [B] (for binary) or [B, C] (for multi-label)
-        """
-        # Convert to probabilities
-        x_sigmoid = torch.sigmoid(x)
-        xs_pos = x_sigmoid
-        xs_neg = 1 - x_sigmoid
-        
-        # Convert labels to one-hot if needed
-        if y.dim() == 1:
-            y_one_hot = F.one_hot(y, num_classes=x.size(1)).float()
-        else:
-            y_one_hot = y.float()
-        
-        # Asymmetric Clipping
-        if self.clip is not None and self.clip > 0:
-            xs_neg = (xs_neg + self.clip).clamp(max=1)
-            
-        # Basic CE calculation
-        los_pos = y_one_hot * torch.log(xs_pos.clamp(min=self.eps))
-        los_neg = (1 - y_one_hot) * torch.log(xs_neg.clamp(min=self.eps))
-        
-        # Asymmetric Focusing
-        if self.gamma_neg > 0 or self.gamma_pos > 0:
-            if self.disable_torch_grad_focal_loss:
-                torch.set_grad_enabled(False)
-            pt0 = xs_pos * y_one_hot
-            pt1 = xs_neg * (1 - y_one_hot)
-            pt = pt0 + pt1
-            one_sided_gamma = self.gamma_pos * y_one_hot + self.gamma_neg * (1 - y_one_hot)
-            one_sided_w = torch.pow(1 - pt, one_sided_gamma)
-            if self.disable_torch_grad_focal_loss:
-                torch.set_grad_enabled(True)
-            los_pos *= one_sided_w
-            los_neg *= one_sided_w
-            
-        loss = los_pos + los_neg
-        return -loss.sum(dim=1).mean()
-
-
-class GradientBasedWeighting:
-    """
-    Gradient-based loss weighting inspired by GradNorm and recent advances
-    """
-    
-    def __init__(self, alpha: float = 1.5, num_tasks: int = 4):
-        self.alpha = alpha
-        self.num_tasks = num_tasks
-        self.initial_losses = None
-        
-    def compute_weights(
-        self,
-        losses: Dict[str, torch.Tensor],
-        shared_params: list,
-        task_specific_params: Dict[str, list]
-    ) -> Dict[str, torch.Tensor]:
-        """
-        Compute gradient-based loss weights
-        """
-        if self.initial_losses is None:
-            self.initial_losses = {k: v.item() for k, v in losses.items()}
-            return {k: torch.ones_like(v) for k, v in losses.items()}
-        
-        # Compute relative loss ratios
-        loss_ratios = {}
-        for task, loss in losses.items():
-            if task in self.initial_losses:
-                loss_ratios[task] = loss.item() / self.initial_losses[task]
-        
-        # Compute gradient norms for each task
-        grad_norms = {}
-        for task, loss in losses.items():
-            if task in task_specific_params:
-                grads = torch.autograd.grad(
-                    loss, shared_params, retain_graph=True, allow_unused=True
-                )
-                grad_norm = torch.sqrt(sum(torch.sum(g**2) for g in grads if g is not None))
-                grad_norms[task] = grad_norm
-        
-        # Compute target gradient norms
-        avg_grad_norm = sum(grad_norms.values()) / len(grad_norms)
-        target_grad_norms = {}
-        
-        for task in grad_norms:
-            if task in loss_ratios:
-                # Target: inverse relationship with loss ratio
-                target_grad_norms[task] = avg_grad_norm * (loss_ratios[task] ** (-self.alpha))
-        
-        # Compute adaptive weights
-        weights = {}
-        for task in grad_norms:
-            if task in target_grad_norms:
-                ratio = target_grad_norms[task] / (grad_norms[task] + 1e-8)
-                weights[task] = torch.clamp(ratio, 0.1, 10.0)
-            else:
-                weights[task] = torch.ones_like(losses[task])
-                
-        return weights
-
-
-# Configuration for different medical scenarios
-MEDICAL_SOTA_CONFIGS = {
-    "rare_disease_detection": {
-        "strategy": "adaptive_distillation",
-        "focal_alpha": 0.1,  # Heavy focus on minority class
-        "focal_gamma": 3.0,   # Strong focusing
-        "ema_decay": 0.9999,  # Very slow EMA for stability
-        "description": "Optimized for rare disease detection with extreme imbalance"
-    },
-    "multi_condition_screening": {
-        "strategy": "gradient_adaptive",
-        "focal_alpha": 0.25,
-        "focal_gamma": 2.0,
-        "ema_decay": 0.999,
-        "description": "Balanced approach for multiple condition screening"
-    },
-    "treatment_response_prediction": {
-        "strategy": "uncertainty_aware",
-        "focal_alpha": 0.3,
-        "focal_gamma": 1.5,
-        "ema_decay": 0.995,
-        "description": "Confidence-aware for treatment response prediction"
-    },
-    "diagnostic_assistance": {
-        "strategy": "ensemble_distillation", 
-        "focal_alpha": 0.2,
-        "focal_gamma": 2.5,
-        "ema_decay": 0.999,
-        "description": "Comprehensive diagnostic assistance with multiple heads"
-    }
-}
-
-
-def get_sota_loss_strategy(medical_scenario: str = "multi_condition_screening", **kwargs):
-    """
-    Factory function to get SOTA loss strategy for specific medical scenarios
-    """
-    if medical_scenario in MEDICAL_SOTA_CONFIGS:
-        config = MEDICAL_SOTA_CONFIGS[medical_scenario].copy()
-        config.update(kwargs)  # Allow custom overrides
-        
-        return SOTALossStrategy(
-            focal_alpha=config['focal_alpha'],
-            focal_gamma=config['focal_gamma'],
-            ema_decay=config['ema_decay'],
-            strategy=config['strategy']
-        )
-    else:
-        raise ValueError(f"Unknown medical scenario: {medical_scenario}")
-
-
-if __name__ == "__main__":
-    print("🏆 State-of-the-Art Loss Strategies for Medical Classification")
-    print("=" * 70)
-    
-    for scenario, config in MEDICAL_SOTA_CONFIGS.items():
-        print(f"\n📋 {scenario.upper()}:")
-        print(f"   Strategy: {config['strategy']}")
-        print(f"   Description: {config['description']}")
-        print(f"   Focal α: {config['focal_alpha']}, γ: {config['focal_gamma']}")
-        print(f"   EMA decay: {config['ema_decay']}")
-    
-    print(f"\n🔬 Advanced Features:")
-    print(f"   ✅ SoftAdapt: Dynamic loss weighting")
-    print(f"   ✅ Asymmetric Loss: Extreme imbalance handling") 
-    print(f"   ✅ Self-Distillation: EMA teacher with momentum")
-    print(f"   ✅ Gradient-based Adaptation: Real-time weight adjustment")
-    print(f"   ✅ Confidence-aware Consistency: Uncertainty weighting")
