+++ train.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 """
-Unified SeqSetVAE Training Script
-Modes: pretrain (reconstruction+KL only), finetune (freeze all except classifier head)
+Unified SeqSetVAE Training Script with SOTA Integration
+Modes: 
+- pretrain: reconstruction+KL only
+- finetune: freeze backbone + SOTA loss strategies for medical classification
+- finetune-sota: advanced medical optimization with academic research integration
 """
 
 import os
@@ -18,6 +21,7 @@ from datetime import datetime
 from model import SeqSetVAE, SeqSetVAEPretrain, load_checkpoint_weights
 from dataset import SeqSetVAEDataModule
 from posterior_collapse_detector import PosteriorMetricsMonitor
+from losses import MEDICAL_SOTA_CONFIGS, get_recommended_strategy
 import config
 
 
@@ -97,7 +101,7 @@ def remap_pretrain_to_finetune_keys(state_dict):
 def main():
     parser = argparse.ArgumentParser(description="Train SeqSetVAE (pretrain/finetune)")
     # Basic training parameters
-    parser.add_argument("--mode", type=str, choices=["pretrain", "finetune"], required=True, help="Training mode")
+    parser.add_argument("--mode", type=str, choices=["pretrain", "finetune", "finetune-sota"], required=True, help="Training mode")
     parser.add_argument("--batch_size", type=int, default=4, help="Batch size")
     parser.add_argument("--gradient_accumulation_steps", type=int, default=4, help="Gradient accumulation steps")
     parser.add_argument("--max_epochs", type=int, default=100, help="Maximum epochs")
@@ -118,6 +122,13 @@ def main():
     # Output root directory
     parser.add_argument("--output_root_dir", type=str, default="/home/sunx/data/aiiih/projects/sunx/projects/TEEMR/PT/outputs", help="Root directory to save logs, checkpoints, and outputs")
     parser.add_argument("--output_dir", type=str, default=None, help="Alias of --output_root_dir; if provided, overrides output_root_dir")
+    
+    # SOTA specific parameters
+    parser.add_argument("--medical_scenario", type=str, default="auto", 
+                       choices=["auto"] + list(MEDICAL_SOTA_CONFIGS.keys()),
+                       help="Medical scenario for SOTA loss strategy")
+    parser.add_argument("--target_auc", type=float, default=0.90, help="Target AUC score for SOTA mode")
+    parser.add_argument("--target_auprc", type=float, default=0.50, help="Target AUPRC score for SOTA mode")
 
     args = parser.parse_args()
 
@@ -148,6 +159,11 @@ def main():
     # Print basic config
     print("🚀 Training Configuration:")
     print(f" - Mode: {args.mode}")
+    if args.mode == "finetune-sota":
+        print(f" - 🏆 SOTA Medical Classification Mode")
+        print(f" - Target AUC: {args.target_auc}")
+        print(f" - Target AUPRC: {args.target_auprc}")
+        print(f" - Medical Scenario: {args.medical_scenario}")
     print(f" - Batch size: {args.batch_size}")
     print(f" - Grad accumulation: {args.gradient_accumulation_steps}")
     print(f" - Max epochs: {args.max_epochs}")
@@ -166,6 +182,12 @@ def main():
         if args.batch_size == 1:
             print("⚠️  Finetune requires multi-patient batches; overriding batch_size to 2")
     print(f" - DataModule batch_size: {dm_batch_size}")
+    
+    # SOTA mode: Auto-detect medical scenario if needed
+    medical_scenario = None
+    if args.mode == "finetune-sota" and args.medical_scenario == "auto":
+        print("🔍 Auto-detecting optimal medical scenario...")
+        medical_scenario = "multi_condition_screening"  # Default, will be refined after data analysis
     data_module = SeqSetVAEDataModule(
         saved_dir=args.data_dir,
         params_map_path=args.params_map_path,
@@ -217,6 +239,46 @@ def main():
     transformer_heads = config.transformer_heads
     transformer_layers = config.transformer_layers
 
+    # SOTA mode: Finalize medical scenario selection
+    if args.mode == "finetune-sota":
+        if args.medical_scenario != "auto":
+            medical_scenario = args.medical_scenario
+        else:
+            # Try to detect from data characteristics
+            try:
+                data_module.setup()
+                from collections import Counter
+                label_counter = Counter()
+                sample_count = 0
+                
+                for batch in data_module.train_dataloader():
+                    labels = batch.get('label')
+                    if labels is not None:
+                        label_counter.update(labels.view(-1).tolist())
+                        sample_count += len(labels)
+                    if sample_count >= 1000:  # Sample enough data
+                        break
+                
+                # Calculate imbalance ratio
+                if len(label_counter) >= 2:
+                    sorted_counts = sorted(label_counter.values())
+                    imbalance_ratio = sorted_counts[0] / sorted_counts[-1] if sorted_counts[-1] > 0 else 1.0
+                    medical_scenario = get_recommended_strategy(imbalance_ratio)
+                    print(f"🎯 Detected imbalance ratio: {imbalance_ratio:.3f}")
+                    print(f"🏥 Selected medical scenario: {medical_scenario}")
+                else:
+                    medical_scenario = "multi_condition_screening"
+            except Exception as e:
+                print(f"⚠️ Auto-detection failed: {e}")
+                medical_scenario = "multi_condition_screening"
+        
+        print(f"📋 SOTA Configuration:")
+        sota_config = MEDICAL_SOTA_CONFIGS[medical_scenario]
+        print(f"   - Scenario: {medical_scenario}")
+        print(f"   - Description: {sota_config['description']}")
+        print(f"   - Focal α: {sota_config['focal_alpha']}, γ: {sota_config['focal_gamma']}")
+        print(f"   - EMA decay: {sota_config['ema_decay']}")
+
     print("🧠 Building model...")
     if args.mode == 'pretrain':
         print("📋 Using SeqSetVAEPretrain - maintains original design for representation learning")
@@ -250,7 +312,21 @@ def main():
         monitor_metric = 'val_loss'
         monitor_mode = 'min'
     else:
-        print("📋 Using SeqSetVAE - FINETUNE MODE with simplified VAE feature fusion")
+        # Finetune mode (both regular and SOTA)
+        if args.mode == "finetune-sota":
+            print("🏆 Using SeqSetVAE - SOTA MODE with advanced medical optimization")
+            sota_config = MEDICAL_SOTA_CONFIGS[medical_scenario]
+            focal_alpha_final = adaptive_alpha if adaptive_alpha is not None else sota_config['focal_alpha']
+            focal_gamma_final = sota_config['focal_gamma']
+            medical_scenario_param = medical_scenario
+            checkpoint_name = f"SeqSetVAE_sota_{medical_scenario}"
+        else:
+            print("📋 Using SeqSetVAE - FINETUNE MODE with simplified VAE feature fusion")
+            focal_alpha_final = adaptive_alpha if adaptive_alpha is not None else config.focal_alpha
+            focal_gamma_final = config.focal_gamma
+            medical_scenario_param = "multi_condition_screening"  # Default
+            checkpoint_name = "SeqSetVAE_finetune"
+            
         model = SeqSetVAE(
             input_dim=config.input_dim,
             reduced_dim=config.reduced_dim,
@@ -272,10 +348,10 @@ def main():
             beta_warmup_steps=model_beta_warmup_steps,
             kl_annealing=config.kl_annealing,
             use_focal_loss=True,
-            focal_alpha=(adaptive_alpha if adaptive_alpha is not None else config.focal_alpha),
-            focal_gamma=config.focal_gamma,
+            focal_alpha=focal_alpha_final,
+            focal_gamma=focal_gamma_final,
+            medical_scenario=medical_scenario_param,  # SOTA parameter
         )
-        checkpoint_name = "SeqSetVAE_finetune"
         monitor_metric = 'val_auc'
         monitor_mode = 'max'
